order model finalized

+++ backend/src/main/java/com/example/e_commerce_backend/dtos/order/OrderDto.java
@@ -1,11 +1,14 @@
 package com.example.e_commerce_backend.dtos.order;
-
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
 //what to return for GET order request
 @Builder
-public record OrderDto(CustomerDto customer, Instant createdAt, Double total, List<OrderItemDto> items) {
+public record OrderDto(CustomerDto customer,
+                       Instant createdAt,
+                       BigDecimal total,
+                       List<OrderItemDto> items) {
 }
